Add reflections about deployment
Deployment was something I found quite challenging - I tried to use aws to deploy the site. However in the time period I allowed to see if it was viable I was unable to make enough progress. So I pulled back and used heroku. I've added reflections and will be looking for ways to develop and improve from this.